wip: hack on substore support
Towards #3129, so far I just started trying to create a `SubstoreSnapshot` that
would do all of the existing `Snapshot` reads on a parameterized set of column
families, by copying over the existing reader code and amending it. Once that's
done, we could replace the existing code with calls to the `SubstoreSnapshot`,
and then write a multistore that would dispatch by prefix.