deploy: make peer limits customizable
We've been raising the peering limits on our PL nodes, as the testnet
grows, maxing out concurrent connections.

We also update the CometBFT default value for `p2p.allow_duplicate_ip`
false -> true. The purpose of this option is raise the cost for
attackers to spam a network with node/validator identities run on the
same host. That's relevant for mainnet, but during testnet, we want to
maximize the ability to peer, so newcomers can join in.

Refs #3056.