build(pd): upgrade to tower-abci v0.10.0
While working on #2263, noticed that we aren't using the most recent
version of tower-abci. This change bumps to latest, ahead of upcoming
more substantive changes in v0.11.0 [0]. Opted to preserve the TCP-only
ABCI support for pd out of simplicity.

Also includes mention of `jmt` in the cargo lockfile as a follow-up to
f482b485ad265511e2040fa2d9090c5bc04273c7.

[0] https://github.com/penumbra-zone/tower-abci/issues/34