feat(pd): special handling of seed nodes in join
When joining a network via `pd testnet join`, pd already inspects
the peers known to the target bootstrap node, and includes those peers
in the generated CometBFT config. Technically it's not correct to assume
that these peers are seeds, but limiting to a threshold (#3063) ensures
that only a few nodes will be miscategorized this way.

The new behavior introduced here is that `pd testnet join` will inspect
monikers of peers learned from the bootstrap node, and if the string
"seed" appears in the moniker, guarantee that peer lands in the `seeds`
field of the CometBFT config.

To complement this change, we update the deployment logic to create a
seed node, and ensure that seed node is excluded from the loadbalanced
RPC, to avoid hitting it via the join request (in which case it won't
report many peers, because seed nodes drop peer connections after
serving address info).

Refs #3056.