ci: fix moon names in host monikers
Follow up to #3155. We've been using moon names as a mnemonic for
human-readable names for services related to Penumbra testnets. Due to a
typo, the moon of Neptune "Naiad" was misspelled, and that typo made it
into a few config files. This is a purely cosmetic change with zero
impact on service behavior or performance.