deploy: update docker-compose setup
Dusted off the compose setup and updated it to use an initcontainer,
same as with the recent overhaul of deploy logic (#2983). This change
also removes the requirement for the host machine to use `pd` to
bootstrap the config: now, docker-compose is all that's required.
The goal is to make the Penumbra containers easier to work with,
for example for the block explorer push.