fix(pd): raise max grpc msg 4MB -> 20MB
Building on the work in summonerd [0], we raise the maximum permitted
gRPC message size, in order to accommodate for initchain ABCI messages
that pass a genesis file of >4MB.

There are side-effects throughout the codebase, mostly related to the
View service, and how it's invoked. To raise max message size, we must
tack on attributes to the ServiceServer and ServiceClient constructions.
Straightforward enough, if non-DRY.

The changes here seem to be necessary but insufficient; I've gotten the
failing smoke tests (for pcli) down from 8 to 2.

TODO - there's a lot of print-based debugging messages, snip those out.