deploy: add support for cometbft postgres indexing
The penumbra-node chart now has opt-in support for storing CometBFT
events in a postgres sidecar. See relevant docs at [0]. The sole
downstream use case for this work to date is cuiloa [1], a block
explorer for Penumbra. There are some shortcuts in this logic, mostly
related to hardcoding db connection info, but the db is intra-cluster
cluster only, so the risk is minimal. Prioritizing developer velocity right now.

[0] https://docs.cometbft.com/v0.37/app-dev/indexing-transactions#postgresql
[1] https://github.com/ejmg/cuiloa