Revert "Summonerd: tie view service failures to crash main service"
This reverts commit bb795bd0dca5de434688107f37005bc1580ede52.

The `wait_for_crash` method doesn't work correctly and will generate new
crashes while waiting to check if the view service crashed.  We need to use the
code we actually did the dry run with.  As per this morning's discussion,
making the view service more robust is something we should work in, but it's
something that requires design work and code review and should not be dropped
only into the summonerd service without testing.