ci: enable expensive rpc service for fullnodes
Here we unconditionally enable serving expensive gRPC services via pd
for the fullnodes. We could make this option configurable in the
deployment logic, but our use case is that we want the expensive options
to be available to the dev team, so just forcing them all on. Notably we
don't expose the RPC services for the genesis validators, and we're not
enabling the flag on pd for the gen vals, so we'll monitor performance
on the load-balanced fullnode endpoints and adjust if necessary.