ci: update cargo-dist for performing releases
In this commit we move from cargo-dist 0.0.4 to 0.4.2.
The tooling updates are considerable, as the version number implies.
To test, I regenerated the workflow YAML and run it many times in a
private fork of the repo. With the notable change that we're no longer
building binaries for Windows (#3298), everything else looks to work
pretty well. Even better, the runtime for the workflow has decreased
from ~60m to ~20m, with the biggest win skipping Windows builds.

Refs #3279, #3298.