fix(pcli): remove archives from debug output
Since we moved to new pcli config file in #3239, we don't use the
archives anymore. If a machine doesn't have archive dirs, then pcli will
throw an error, which is bad. Stop looking for archive dirs.

Similar, we shouldn't look for a custody.json file, we should look for
the new config.toml file.

Closes #3321.