tx: simplify EffectHash computation
I'd missed that this change wasn't already included in #2808. I thought that
we'd moved to the simpler alternative of making the `EffectHash` just be the
hash of all of the actions in the plan, and expecting the transaction builder
to build the actions in the order they appear in the plan.

Not doing this totally blows up the complexity of the changes we're trying to
achieve now.  However, making this change means that we can simply build the
actions in the order they appear in the plan and get correct results.