feat(pd): custom voting periods testnet generate
During testing of upgrades, it's been handy to reduce the default,
static value of "24h" for voting on a given proposal, so that a proposal
can be made, voted on, and passed, all within e.g. 1h. This setting
greatly accelerates testing on devnets.

Also includes customization var in deploy scripts for `epoch_duration`,
which we're not currently overriding, but now we can if we want to.

Ideally we'd set these customizations by just overriding the
`pd testnet generate` command, but given that the cmd requires
a storage dir arg that's non-standard, it's more straightforward
to stick with the clunky vars-based approach for now.