deploy: increase node storage 100G -> 200G
After five weeks running Testnet 63 Rhea (#3182), the PL-run nodes
filled up their available disk storage, which is 100G per node. (Of
that, roughly 90% is pd, 10% cometbft.) To recover, we manually
edited the PVCs to grow to 200G, and this change locks that new config
into version control, for use in future testnets.