pd: allow bundling the web frontend into the fullnode
This is an experiment towards bundling the web frontend code into the fullnode,
with the goal that users who run `pd` nodes can self-host a complete Penumbra
frontend, on their custom domain with auto-HTTPS, with just one command: `pd start`.

To do this, we

- add a new `zipserve` module in `pd`, which takes a `&'static [u8]` with archive data and produces an `axum::Router`;
- bundle an `assets/frontend.zip` (tracked with git-lfs) into the `pd` crate;
- use `axum::Router::merge` to combine the GRPC and frontend endpoints.

To make this work for real, we'd need to:

- [ ] Fix up the auto-HTTPS support so that it can work directly with Axum, rather than Tonic;
- [ ] Change the frontend code so that all URLs are relativizable to an arbitrary prefix (we want to have it at `/app` when served in `pd`, to allow a status page, but `/` when at `app.testnet.penumbra.zone` or other deployments);
- [ ] Create a process to regenerate `assets/frontend.zip` from the `web` repo periodically and check the result into the repo (under LFS).

These seem pretty doable, and it'd be really cool to be able to have easily self-hosted frontends!