feat(pd): report bind errors on startup
A common problem for node operators is sorting out which ports are
required to be available on the system prior to running. We've recently
improved the documentation to be more explicit about this [0], but here
we add explicit error messages when binds fail. This addition should
make reasoning about network surface more straightforward, and eliminate
guessing games to figure out which port needs to be overridden or freed.

[0] c0445f34752f64a253b672188f3a6d8f5f58c116