ci: add moons of mars to testnet names
We've shipped testnets for the Jovian and Saturnine moons.
Now, we add the Mars of moons, only two (2). If we need more than this,
we can follow up with the moons of Neptune for more headroom.

Also updates the names of fullnodes in the CI deployment, which were
previously set to moons of Mars and Neptune. Now moon names are only
used to refer to testnets, and the fullnode monikers are notable
asteroids.