ci: disable osiris handling
We've disabled Osiris, likely indefinitely, so let's remove it from the
CI workflows. Otherwise, preview deploys report as failed, even though
the preview environment was created succesfully. In the medium term, we
hope to have a Hummingbot service running, which will serve a similar
role to Osiris.

Also removes the unused relayer reboot, as we're using Hermes for IBC
now.