ibc: ⛅ place validate_penumbra_client_state in ics02_validation
nb: this commit strictly contains plain code motion. it does not make any
changes to the code being moved.

while this was not originally placed in the ics02_validation submodule,
it feels like it is more connected to that validation logic than the
surrounding client counter types.

this bears out to be true; it is `pub`, the client counter code isn't
changed, and affected imports are in the message handler area.