fee: fix regression in #3624
In #3624, the default gas prices were changed to have nonzero values, in order
to test client functionality with nonzero fees. However, we *don't* want to
make this the default until we've ported all the client software over, and even
then, we should make the gas price setting be part of the `AppParameter`s,
rather than changing the default value. (We want the default value for
`GasPrices` to align with the proto default value).