feat(pcli) default to phase 2 for ceremony
As plan to transition to phase 2 of the summoning ceremony along with
the release of Testnet 65. Here we update the default value for
`pcli ceremony contribute --phase` to be `2`, rather than a required
arg. The summoning website already displays an explicit example command,
and the value for "phase" is determined by the phase of the coordinator
server, so no changes needed there.