distributions: issuance budget in staking units
Previously, the distribution component would issue a budget
that was _not_ denominated in the staking token unit. This means
that the staking component would have to do some ad-hoc conversion
in order to use it. This is confusing (in fact, it caused a bug).

Instead of doing that, we just track the issuance budget and apply
the staking token denom unit amount.