staking(rate): rework `RateData::delegation_amount`
Part of an effort to systematize how we perform rate calculations,
hopefully making things a little less hairy for the next person.

First, we strip the scaling factors. This gives us rates between
zero and one. Those rates are applied to `Amount`s of staking tokens
or delegation pool tokens.

If the calculation is a rate itself (e.g. `RateData::next_epoch`) we
re-apply the scaling factor (`FP_SCALING_FACTOR` i.e. 1e8) so that
we can store the rate in an `Amount` in our public interfaces.