staking(rate): adjust end-epoch processing
I have add many noisy debug statements as part of my testing phase.

Recording it here for future reference (possibly). Something that
came up up during testing is that a good way to sanity check rate
calculations is to spin-up a testnet with a single validator with
commission 100% and a single funding stream. Then, we tune the epoch
duration and make sure that no bad conditions are hit (e.g. divide
by zero), and that the amount of reward piped to the validator
corresponds exactly to the issuance budget for that epoch.

This hints to a testing structure we could adopt.