staking(funding): fix funding stream calculations
This was a holdover from a previous time-based mechanism push. Previously, rewards were computed on the first block of a new epoch. This meant that to compute the rewards for the previous epoch, one had to use the corresponding old rate data. This is why this method was taking both the previous and current rate data as parameters.

Moreover,