ci: fix smoke test
There were a few problems with the changes in #3656. First, the YAML
workflow file was poorly indented, which caused the job not to be
evaluated, so we got a false green.

Beyond that, the wrapper script needed handling for an undefined var,
and a tweak to the backgrounding/logging changes. All are updated, and
this change passes locally.