shielded-pool: remove `update_token_supply`
In #3640, we added two helper methods:
- `increase_token_supply`
- `decrease_token_supply`

These methods are somewhat similar, and objectively are more
verbose than a single `update_token_supply` method. However,
it seems an improvement to not have to do any cognitive work
to piece together whether a method call increases or decreases
the token supply.

Another advantage of this approach is that one can easily use
rust-analyzer to filter for places that increase the token supply
vs. places where it is being shrinked.