penumbra-ibc: ⏎  invert early return condition
this diff is best viewed ignoring whitespace; use `-w`, or the "_hide
whitespace_" option in github (the gear icon in the "_files changed_"
tab).

this inverts our early return, so that we do not need to wrap the whole
function body within a conditional block. this is a pure refactor that
should have no visible changes to behavior.