ci: workflow for standalone pd
We want to exercise the pd https logic, but we can't naively run it
from scratch on every deploy, because that'd be far too many API
requests to reissue certs from ACME. Instead, let's preserve the ACME
directory before wiping state, and reuse it before bouncing the service.

This setup requires always-on boxes provisioned out of band.
So far, this adds the base logic via a workflow. In order to get it
running, I'll need to iterate on the workflow, but workflows must land
on main prior to being available for ad-hoc execution.

Refs #3336.