asset: allow image URIs in asset metadata
This commit expands the `DenomMetadata` proto message to support the `images`
field defined in the cosmos asset registry extensions.

Per discussion on Telegram, the `logo_URIs` and `images` fields overlap in
functionality, but `images` is more forward-compatible and extensible. It also
has a proto-compatible name, so seems like a win to only use it.

It also adds a test that we can parse metadata defined in the registry. To
enable this, we need to configure pbjson-build to ignore unknown fields, which
is the normal proto behavior.