sct(component): add `component::clock`
I am not completely sold on the name, but the idea is that what you
find here are extension traits that let you manage the blockchain
clocks. Those clocks can be logical e.g. a block or an epoch, or
literal wall clock (recorded by block timestamps, and eventually
PBTS).

In particular, we define `EpochManager` and `EpochRead` which
implement write and read capabiltiies respectively.