sct(component): add `component::Tree`
In this module, we find composable extension traits that let us
interface with the chain's commitment tree. This is done by plugging
into `SctRead` or `SctManager`.

We also add a third interface (`tree:VerificationExt`) which action
handlers can pull in order to perform SCT specific verification of
anchors or nullifiers.