sct(component): implement `Component` trait and split modules  (#3723)
* sct(view): remove catch-all view module

* sct(component): remove top-level module sig

* sct(component): add `component::clock`

I am not completely sold on the name, but the idea is that what you
find here are extension traits that let you manage the blockchain
clocks. Those clocks can be logical e.g. a block or an epoch, or
literal wall clock (recorded by block timestamps, and eventually
PBTS).

In particular, we define `EpochManager` and `EpochRead` which
implement write and read capabiltiies respectively.

* sct(component): add `component::Tree`

In this module, we find composable extension traits that let us
interface with the chain's commitment tree. This is done by plugging
into `SctRead` or `SctManager`.

We also add a third interface (`tree:VerificationExt`) which action
handlers can pull in order to perform SCT specific verification of
anchors or nullifiers.

* sct(component): move `SourceContext`

This is a purely mechanical refactor, and might be better to
merge into the tree module to increase cohesion.

* sct(component): update query service imports

* sct(component): implement `cnidarium_component::Component` on the sct

* sct(component): propagate api changes

* sct(component): sketch out updated state key schema

* app: use `Sct::init_chain` and `Sct::begin_block`

* sct(component): track module definition

* sct(component): use updated state key schema

* penumbra: misc small fixes