staking(validator_manager): implement multi-tombstoning
When a validator commits byzantine misbehavior, its delegation pool
is slashed, unbonded, and its state is set to a terminal `Tombstoned`
value. Previously, we assumed that an offender could not recidive
and considered double `Tombstoning` to be an error.

But that's not correct, for example a validator could have been
equivocating at two different heights, or rounds. Similarly, in
theory, it could perform various kind of byzantine misbehavior (see
`MisbehaviorKind`).