staking: make multi-tombstoning a no-op and a state machine fix (#3761)
* staking(validator_store): rename `ValidatorStore` -> `ValidatorDataWrite`

* staking(epoch_handler): propagate api change

* staking(validator_manager): add simple `Disabled <> Defined` transition

This commit makes double-tombstoning a no-op (for now), as opposed
to an error. And it implements the `Defined <> Disabled` state
transitions.

* staking(validator_manager): move `Active -> Jailed` transition up

* staking(validator_manager): improve comments on state machine

* staking: make `get_penalty_in_epoch` infaillible

* staking(validator_manager): make `Active -> Jailed` more visible

* staking(validator_manager): implement multi-tombstoning

When a validator commits byzantine misbehavior, its delegation pool
is slashed, unbonded, and its state is set to a terminal `Tombstoned`
value. Previously, we assumed that an offender could not recidive
and considered double `Tombstoning` to be an error.

But that's not correct, for example a validator could have been
equivocating at two different heights, or rounds. Similarly, in
theory, it could perform various kind of byzantine misbehavior (see
`MisbehaviorKind`).

* staking: make Id(Tombstoning) a no-op