chore(flake): 🙄 use rustup instead of overlay
when new versions of rust are released, there is a period before the new
version appears in the overlay. this, in the worst case, can leave me (a
foolish nix user) with a different toolchain than used in CI, with no
way to easily rectify this.

we could (should?) use a toolchain file, but i'm not going to introduce
one purely for nix's sake. eugh.