chore: update chain id for 67 phobos
Generates a new chain id for preview, for the 67 phobos chain.
Also copy/pastes the 65 allos to 66, see #3779 for context.

Also added a new address for allocations for Park.

Also squashed in:

fix: remove test address from base allocations

The test seed phrase was declared twice, once in test-addresses, which
is good, and again in the "base allocations", which are too large.