proto: allow more information in ValueView
This attempts to make the ValueView more useful for frontends, and prevent
frontend code from having to invent its own handling of numeraires.  Here, we
make two extensions to the ValueView:

1. Allow adding a list of "equivalent values" in different numeraires. We want
to have support for multiple equivalent values because it may be useful to
relate the viewed value to multiple different assets. For instance, a
delegation token has at least two relevant numeraires: how many staking tokens
it represents (using the per-validator exchange rate) and what its cash value
is (in, say, USDC).

2. Allow adding arbitrary extended data using an `Any`. This could be used to
insert a `component.dex.Position` with the complete position state into the
`ValueView` of an LPNFT, allowing frontend code that has knowledge of LPNFTs to
show a rendering of the position's current reserves, price, etc. Using an `Any`
provides extensibility and graceful degradation.

These extensions are not added to the Rust domain types yet, since they're
intended for the web context.