chore: create "deimos-2" chain id
Testnet 66 experienced a chain halt. To recover, we're going to release
a new chain id, "penumbra-testnet-deimos-2", in an old-school
nuke-and-redeploy scenario. The team discussed the opportunity to
exercise chain upgrade logic, but decided it'd be better to prepare docs
for validator operations ahead of time.

This change effectively reverts (or moots) 8c5cb5b8e088ca585c724c7c7f57a1f0f1dda997,
which prematurely bumped the chain id on main to phobos. We're not using the phobos name yet,
because the 65, 66, and now 67 versions are logically part of the deimos
series.

Ideally we'd have a more ergonomic way of incrementing the chain
revision number, but we'll come back to that as part of a general
refactor.