ci: add indexing node for dex-explorer
We're starting work on a new webapp that will provide a visual frontend
for the DEX. That work is currently housed in [0]. Similar to cuiloa
[1], we must enable CometBFT indexing to Postgres, so the application
(yet to be written) can query those events. Strictly speaking, it's
possible to reuse a db connection across multiple apps, particularly if
we're careful about ensuring read-only access, but duplicating the node
config per-app will simplify debugging in the near-term.

These changes have been manually applied for both environments. URLs may
change in the future.

[0] https://github.com/penumbra-zone/dex-explorer
[1] https://github.com/penumbra-zone/cuiloa