governance: 🎾 `DelegatorVoteProof::verify` is fallible (#3833)
see https://github.com/penumbra-zone/penumbra/issues/3777.

this changes `DelegatorVoteProof::verify` so that it is fallible. errors
will be propagated to the caller, rather than inducing a panic.

* #3777
* #3829

as with #3829, i've applied some tlc to the code as i've addressed the
core issue at hand (_panics_). noop code transformations are applied in
distinct commits, to facilitate ease of review.