mock-consensus: 🙂 simplify test node startup (#3849)
this is a bit of polish for the mock consensus tests. `Consensus`
provides its `tower::Service` implementation via `tower-actor`, which
makes naming a `TestNode<_>` slightly verbose. this introduces a
`ConsensusService` type alias, and outlines our test node startup into a
shared function in `common`.

in the future we'll probably extend this `start_test_node` to allow for
callers to override particular settings on the test node builder. for
now, that is elided. our primary goal is to provide an ergonomic "_start
a test node backed by temporary storage_" button to reduce test
boilerplate.

* #3588
* #3788