feat: penumbra-network crate
Ports the network-related code out of `pd` into a new crate
`penumbra-network`. The core functionality we care about boils down to
the existing pd subcommands:

  * pd testnet join
  * pd testnet generate

In the process, overhauls how config logic for fullnode, validator,
and genesis are constructed. The goal is to make the code easier to
reason about, and much easier to reuse, e.g. in testing harnesses.