ci: enable ratelimiting via traefik
As a temporary measure, we've added ratelimiting on the gRPC endpoint
for pd, via Traefik middleware. We plan to add similar ratelimiting
functionality to pd itself, so that solo node operators can benefit from
limited DoS protection. For now, we add this so that a subsequent CI
deploy (e.g. #3784) doesn't remove the protection that's currently live.