staking: remove `component::FP_SCALING_FACTOR` dup (#3895)
This should fix the wasm build unless something else is wrong. The
problem was that we defined the same constant (`FP_SCALING_FACTOR`)
twice, once correctly at the crate-level, and another one in
`penumbra_stake::component`. During development, the latter one was used
and created a discreet dependency on the component module.

x-ref #3894

---------

Co-authored-by: turbocrime <turbocrime@users.noreply.github.com>