dex: avoid chain halts in case of arb failures
In #3893 we encountered a bug in arb search logic that caused a chain halt.
While that bug shouldn't exist, it also shouldn't have stopped the entire
chain. Rather, we should continue operating without performing arbitrage.

This commit does not fix the bug in #3893, but it does mean that the effect of
that bug is just to stop arb from occurring rather than halting the chain.