asset: add proto and domain support for equivalent values
I added the `EquivalentPrice` to the asset proto package rather than solely in
the TxP, because I realized that the use cases we care about most immediately
(having equivalent prices for balances) don't actually involve
`TransactionView`s at all.

While doing that I realized that in the context of viewing a _transaction_, it
doesn't really make sense to just have an "equivalent value" on its own, so I
added an `as_of_height` to allow indicating when the prices were relevant.  We
probably shouldn't try to expose this in the transaction view just yet, since
it's a little unclear how we'd want to indicate historical prices.