refactor(pd): ⭐ hoist grpc server out of `main`
code motion, part 1 of 2.

see #3913. we want to run a view client in application tests, so that we
can use the `Planner` to build transactions. in order to create a view
client, we need to run a grpc server that it can talk to.

NB: we leave the "expensive" simulation service and the tendermint proxy
out of this function for now.