grpc-server: 🍣 move constructor out to crate
and lo, we move the server out to the crate we made!

this will let us spin up the grpc server for use in penumbra-app tests,
so that we may run the view service and work with the Planner API.

this has another pleasant side effect, in slimming down the dependency
tree of `pd` rather substantively.