tests(pcli): 🔇 make large tracing event trace-level (#3989)
there are two issues at hand:
- we log `undelegation_result` twice
- `undelegation_result` can be many thousands of lines long

this means that the `delegate_and_undelegate` pcli network integration
test thrashes the test output somewhat severely, and test logs aren't
fully viewable in e.g. tmux, when running this locally.


https://github.com/penumbra-zone/penumbra/actions/runs/8210700115/job/22458602964#step:11:18

as of `main` at the time of writing, this CI job logs an event that
is...

```
; xclip -o -selection clipboard | wc --bytes
243775
```

243Kb is disruptively long. (_let alone 486Kb!_)

so, let's make this field an opt-in field to display. we only log it
once, and do so at the trace level.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label.