mock-consensus: 🔀 test node can `fast_forward`
fixes #3933.

this introduces a `fast_forward` method to a test node, allowing tests
using the mock consensus engine (#3588) to fast forward a certain number
of blocks.

the existing `mock_consensus_can_send_a_sequence_of_empty_blocks` test
is rewritten to make use of this method.

this is done in service of #3995.

---

* #3588
* #3933
* #3995